Update animatedprofile.py
1.0.0

+++ animatedprofile.py
@@ -34,17 +34,19 @@ logger = logging.getLogger(__name__)
 # сам класс модуля
 @loader.tds
 class AnimatedProfile(loader.Module):
-    """Module for your profile animation (name, bio) | Модуль для анимации вашего профиля (имя, био)"""
+    """Module for your profile animation (name, bio) look in the config | Модуль для анимации вашего профиля (имя, био) смотрите конфиг"""
     strings = {
         "name" : "AnimatedProfile",
         "name_delay": "Time between frames of name animation",
         "animated_name_frames": "Name animation frames",
+        "not_name_frames": "<emoji document_id=5447644880824181073>⚠️</emoji> See the config! In the animated_name_frames parameter, put your animation frames by name",
         "name_is_enabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Name animation is already enabled, use <code>.astopname</code> to turn it off.",
         "name_is_disabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Name animation is already turned off.",
         "name_turned_off": "<emoji document_id=5447644880824181073>⚠️</emoji> Name animation is disabled.",
         "bio_status": "Is the bio animation enabled or not",
         "bio_delay": "Time between frames of bio animation",
         "animated_bio_frames": "Bio animation frames",
+        "not_bio_frames": "<emoji document_id=5447644880824181073>⚠️</emoji> See the config! In the animated_bio_frames parameter, put your animation frames bio",
         "bio_is_enabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Bio animation is already enabled, use <code>.astopname</code> to turn it off.",
         "bio_is_disabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Bio animation is already turned off.",
         "bio_turned_off": "<emoji document_id=5447644880824181073>⚠️</emoji> Bio animation is disabled.",
@@ -52,12 +54,14 @@ class AnimatedProfile(loader.Module):
     strings_ru = {
         "name_delay": "Время между кадрами анимации имени",
         "animated_name_frames": "Кадры анимации имени",
+        "not_name_frames": "<emoji document_id=5447644880824181073>⚠️</emoji> Смотрите конфиг! В параметре animated_name_frames, поставьте ваши кадры анимации имени",
         "name_is_enabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация имени уже включена, используйте <code>.stopaname</code>, чтобы выключить.",
         "name_is_disabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация имени уже выключена.",
         "name_turned_off": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация имени выключена.",
         "bio_status": "Включена ли анимация био или нет",
         "bio_delay": "Время между кадрами анимации био",
         "animated_bio_frames": "Кадры анимации био",
+        "not_bio_frames": "<emoji document_id=5447644880824181073>⚠️</emoji> Смотрите конфиг! В параметре animated_bio_frames, поставьте ваши кадры анимации био",
         "bio_is_enabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация био уже включена, используйте <code>.stopabio</code>, чтобы выключить.",
         "bio_is_disabled": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация био уже выключена.",
         "bio_turned_off": "<emoji document_id=5447644880824181073>⚠️</emoji> Анимация био выключена.",
@@ -96,6 +100,8 @@ class AnimatedProfile(loader.Module):
     @loader.command(alias="aname", ru_doc="""(aname) Включить анимацию имени""")
     async def animatedname(self, message: Message):
         """(aname) Turn on name animation"""
+        if self.config["animated_name_frames"] == []:
+            return await utils.answer(message, self.strings("not_name_frames"))
         if self.aname is False:
             self.aname = True
             await message.delete()
@@ -110,6 +116,8 @@ class AnimatedProfile(loader.Module):
     @loader.command(alias="abio", ru_doc="""(abio) Включить анимацию био""")
     async def animatedbio(self, message: Message):
         """(abio) Turn on bio animation"""
+        if self.config["animated_bio_frames"] == []:
+            return await utils.answer(message, self.strings("not_bio_frames"))
         if self.abio is False:
             self.abio = True
             await message.delete()
